增加AOStorageService用于map合并和切换
