+++ lealone-storage/src/main/java/org/lealone/storage/BufferedMap.java
@@ -265,6 +265,7 @@ public class BufferedMap<K, V> implements StorageMap<K, V>, Callable<Void> {
     public void close() {
         buffer.clear();
         map.close();
+        AOStorageService.removeBufferedMap(this);
     }
 
     @Override
